+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Added CloudInjectionModule to cloud-velocity
  - Added PlayerArgument to cloud-velocity
  - Added TextColorArgument to minecraft-extras
+ - Added ServerArgument to cloud-velocity
 
 ## [1.0.1] - 2020-10-14